+++ app/components/url/url-controller.js
@@ -24,7 +24,6 @@ angular.module('lizard-nxt')
   'MapService',
   'NxtRegionsLayer',
   'State',
-  '$rootScope',
   'LeafletService',
   'gettextCatalog',
   function (
@@ -39,7 +38,6 @@ angular.module('lizard-nxt')
     MapService,
     NxtRegionsLayer,
     State,
-    $rootScope,
     LeafletService,
     gettextCatalog
   ) {
@@ -234,7 +232,8 @@ angular.module('lizard-nxt')
      */
     $scope.$watch(State.toString('context'), function (n, old) {
       if (n === old) { return true; }
-      state.context.update = false;
+      state.context.update = true;
+
       LocationGetterSetter.setUrlValue(
         state.context.part, state.context.index, State.context
       );
@@ -272,7 +271,6 @@ angular.module('lizard-nxt')
         LocationGetterSetter.setUrlValue(state.context.part, state.context.index, state.context.value);
         $scope.transitionToContext(state.context.value);
       }
-      $rootScope.context = State.context;
 
       if (boxType) {
         State.box.type = boxType;